+++ cubedpandas/__init__.py
@@ -2,7 +2,7 @@
 
 import cubedpandas.context as context
 from cubedpandas.common import cubed
-from cubedpandas.datetext.parser import parse, Parser
+from cubedpandas.datefilter.date_filter import DateFilter
 from cubedpandas.ambiguities import Ambiguities
 from cubedpandas.context.context import Context
 from cubedpandas.context.context_context import ContextContext
@@ -53,8 +53,7 @@ __all__ = [
     "MemberContext",
     "MemberNotFoundContext",
     "MemberSet",
-    "parse",
-    "Parser",
+    "DateFilter",
     "Schema",
     "VERSION"
 ]