+++ tests/test_datetext_parser.py
@@ -1,8 +1,12 @@
 from unittest import TestCase
 from datetime import date, datetime
+
+import numpy as np
+import pandas as pd
 from dateutil.parser import parse
-from cubedpandas.datetext import parse
-from cubedpandas.datetext.tokenizer import Tokenizer, Token, TokenType
+
+from cubedpandas import DateFilter
+from cubedpandas.datefilter.tokenizer import Tokenizer, Token, TokenType
 
 class TestDateTextParser(TestCase):
     def setUp(self):
@@ -23,14 +27,14 @@ class TestDateTextParser(TestCase):
 
     def test_initial_parse_using_dateutil(self):
         text = "2024-09-09"
-        result = parse(text)
-        self.assertEqual(result[0], datetime(2024, 9, 9))
+        result = DateFilter(text)
+        self.assertEqual(result[0][0], datetime(2024, 9, 9))
 
     def test_single_word(self):
         text = "today"
-        result = parse(text)
-        self.assertEqual(result[0], datetime.now().replace(hour=0, minute=0, second=0, microsecond=0))
-        self.assertEqual(result[1], datetime.now().replace(hour=23, minute=59, second=59, microsecond=999999))
+        result = DateFilter(text)
+        self.assertEqual(result[0][0], datetime.now().replace(hour=0, minute=0, second=0, microsecond=0))
+        self.assertEqual(result[0][1], datetime.now().replace(hour=23, minute=59, second=59, microsecond=999999))
 
     def test_single_word_month(self):
         words = ["today", "yesterday", "tomorrow", "week", "month", "quarter", "year", "ytd", "mtd", "qtd", "wtd",
@@ -39,9 +43,30 @@ class TestDateTextParser(TestCase):
                  "july", "august", "september", "october", "november", "december",
                  ]
         for word in words:
-            result = parse(word)
+            result = DateFilter(word)
             if self.debug:
-                print(f"{word} = {result[0]} - {result[1]}")
+                print(f"DateFilter: {word}:")
+                for i, span in enumerate(result):
+                    print(f"\t{i + 1:02d} = {span[0]} - {span[1]}")
+                print(f"\tSQL: {result.to_sql('order date')}")
+
+    def test_to_lambda_sql_function(self):
+
+        df = DateFilter("today")
+        sql = df.to_sql("order date")
+        f_func = df.to_function()
+        f_lambda = df.to_lambda()
+        f_numpy = df.to_numpy_lambda()
+        n_array = np.array([datetime.now() for _ in range(100)], dtype=np.datetime64)
+        df = pd.DataFrame(n_array, columns=["order date"])
+
+        self.assertEqual(f_func(datetime.now()), True)
+        self.assertEqual(f_lambda(datetime.now()), True)
+        result = f_numpy(df["order date"])
+        self.assertEqual(result.index.to_list(), [i for i in range(100)])
+
+
+
 
     def test_tokenizer(self):
         datetexts = ["1st of January 2024",